# docs/admin-api/admin_swagger.yaml
basePath: /api/v1
definitions:
  api.PageData:
    properties:
      list: {}
      pageNum:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  api.Response:
    properties:
      code:
        type: integer
      data: {}
      detail:
        type: string
      msg:
        type: string
      requestId:
        type: string
    type: object
...
  dto.SysDictAddReq:
    properties:
      description:
        type: string
      label:
        type: string
      name:
        type: string
    type: object
  dto.SysDictDeleteReq:
    properties:
      id:
        type: integer
    type: object
  dto.SysDictGetItemsReq:
    properties:
      name:
        type: string
    type: object
  dto.SysDictGetItemsResp:
    properties:
      list:
        items:
          $ref: '#/definitions/models.SysDictItem'
        type: array
    type: object
  dto.SysDictItemAddReq:
    properties:
      description:
        type: string
      dictName:
        type: string
      isDefault:
        type: boolean
      itemLabel:
        type: string
      itemValue:
        type: string
      sort:
        type: integer
    type: object
  dto.SysDictItemUpdateReq:
    properties:
      description:
        type: string
      id:
        type: integer
      isDefault:
        type: boolean
      itemLabel:
        type: string
      itemValue:
        type: string
      sort:
        type: integer
      status:
        type: integer
    type: object
  dto.SysDictPageQueryReq:
    properties:
      idOrder:
        type: string
      keyword:
        type: string
      pageNum:
        type: integer
      pageSize:
        type: integer
    type: object
  dto.SysDictUpdateReq:
    properties:
      description:
        type: string
      id:
        type: integer
    type: object
  dto.SysPermissionAddReq:
    properties:
      anon:
        type: boolean
      auth:
        type: boolean
      description:
        type: string
      name:
        type: string
      parentId:
        example: "0"
        type: string
      type:
        type: string
    type: object
  dto.SysPermissionBatchAddReq:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.SysPermissionAddReq'
        type: array
    type: object
  dto.SysPermissionDeleteReq:
    properties:
      id:
        example: "0"
        type: string
    type: object
  dto.SysPermissionImportReq:
    properties:
      anon:
        type: boolean
      auth:
        type: boolean
      children:
        items:
          $ref: '#/definitions/dto.SysPermissionImportReq'
        type: array
      description:
        type: string
      name:
        type: string
      parentId:
        example: "0"
        type: string
      type:
        type: string
    type: object
  dto.SysPermissionPageQueryReq:
    properties:
      anon:
        type: boolean
      auth:
        type: boolean
      idOrder:
        type: string
      keyword:
        type: string
      pageNum:
        type: integer
      pageSize:
        type: integer
      type:
        type: string
    type: object
  dto.SysPermissionTree:
    properties:
      anon:
        type: boolean
      auth:
        type: boolean
      children:
        items:
          $ref: '#/definitions/dto.SysPermissionTree'
        type: array
      createdAt:
        type: string
      createdBy:
        example: "0"
        type: string
      description:
        type: string
      id:
        example: "0"
        type: string
      key:
        example: "0"
        type: string
      name:
        type: string
      parentId:
        type: integer
      title:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      updatedBy:
        example: "0"
        type: string
    type: object
  dto.SysPermissionUpdateReq:
    properties:
      anon:
        type: boolean
      auth:
        type: boolean
      description:
        type: string
      id:
        example: "0"
        type: string
      name:
        type: string
      parentId:
        example: "0"
        type: string
      type:
        type: string
    type: object
...
  models.SysDict:
    properties:
      createdAt:
        type: string
      createdBy:
        example: "0"
        type: string
      del:
        type: boolean
      deletedAt:
        type: string
      deletedBy:
        example: "0"
        type: string
      description:
        type: string
      id:
        type: integer
      label:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      updatedBy:
        example: "0"
        type: string
    type: object
  models.SysDictItem:
    properties:
      createdAt:
        type: string
      createdBy:
        example: "0"
        type: string
      description:
        type: string
      dictName:
        type: string
      id:
        type: integer
      isDefault:
        type: boolean
      itemLabel:
        type: string
      itemValue:
        type: string
      sort:
        type: integer
      status:
        type: integer
      updatedAt:
        type: string
      updatedBy:
        example: "0"
        type: string
    type: object
  models.SysPermission:
    properties:
      anon:
        type: boolean
      auth:
        type: boolean
      createdAt:
        type: string
      createdBy:
        example: "0"
        type: string
      description:
        type: string
      id:
        example: "0"
        type: string
      name:
        type: string
      parentId:
        type: integer
      type:
        type: string
      updatedAt:
        type: string
      updatedBy:
        example: "0"
        type: string
    type: object
...
host: 127.0.0.1:10000
info:
  contact:
    email: xxx@qq.com
    name: xxx
    url: http://xxx.com
  description: xxx管理后台API文档
  title: Admin API
  version: 1.0.0
paths:
...
  /sys/dict/add:
    post:
      consumes:
      - application/json
      description: 添加字典
      parameters:
      - description: 字典信息
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysDictAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 添加字典
      tags:
      - 字典管理
  /sys/dict/delete:
    post:
      consumes:
      - application/json
      description: 根据id删除字典
      parameters:
      - description: 要删除的字典ID
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysDictDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 删除字典
      tags:
      - 字典管理
  /sys/dict/item/add:
    post:
      consumes:
      - application/json
      description: 添加字典项
      parameters:
      - description: 字典项
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/dto.SysDictItemAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 添加字典项
      tags:
      - 字典管理
  /sys/dict/item/batch_get_by_names:
    post:
      consumes:
      - application/json
      description: 根据字典名称数组获取所有项目
      parameters:
      - description: 字典名称数组
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/dto.SysDictGetItemsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  additionalProperties:
                    items:
                      $ref: '#/definitions/models.SysDictItem'
                    type: array
                  type: object
              type: object
      security:
      - RequireLogin: []
      summary: 根据字典名称数组获取所有项目
      tags:
      - 字典管理
  /sys/dict/item/get_by_name:
    post:
      consumes:
      - application/json
      description: 根据字典名称获取所有项目
      parameters:
      - description: 字典名称
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/dto.SysDictGetItemsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SysDictGetItemsResp'
              type: object
      security:
      - RequireLogin: []
      summary: 根据字典名称获取所有项目
      tags:
      - 字典管理
  /sys/dict/item/page_query:
    post:
      consumes:
      - application/json
      description: 分页查询字典
      parameters:
      - description: 字典筛选条件
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysDictPageQueryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/api.PageData'
                  - properties:
                      List:
                        $ref: '#/definitions/models.SysDict'
                    type: object
              type: object
      security:
      - RequireLogin: []
      summary: 分页查询字典
      tags:
      - 字典管理
  /sys/dict/item/update:
    post:
      consumes:
      - application/json
      description: 修改字典项
      parameters:
      - description: 字典项
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/dto.SysDictItemUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 修改字典项
      tags:
      - 字典管理
  /sys/dict/page_query:
    post:
      consumes:
      - application/json
      description: 分页查询字典
      parameters:
      - description: 字典筛选条件
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysDictPageQueryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/api.PageData'
                  - properties:
                      List:
                        $ref: '#/definitions/models.SysDict'
                    type: object
              type: object
      security:
      - RequireLogin: []
      summary: 分页查询字典
      tags:
      - 字典管理
  /sys/dict/update:
    post:
      consumes:
      - application/json
      description: 根据id修改字典信息
      parameters:
      - description: 字典信息
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysDictUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 修改字典
      tags:
      - 字典管理
  /sys/permission/add:
    post:
      consumes:
      - application/json
      description: 添加权限
      parameters:
      - description: 权限信息
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysPermissionAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 添加权限
      tags:
      - 权限管理
  /sys/permission/batch_add:
    post:
      consumes:
      - application/json
      description: 批量添加权限
      parameters:
      - description: 权限筛选条件
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysPermissionBatchAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 批量添加权限
      tags:
      - 权限管理
  /sys/permission/batch_import:
    post:
      consumes:
      - application/json
      description: 批量导入权限
      parameters:
      - description: 权限树
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysPermissionImportReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 批量导入权限
      tags:
      - 权限管理
  /sys/permission/bind_users:
    post:
      consumes:
      - application/json
      description: 以树形结构呈现所有权限
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SysPermissionTree'
                  type: array
              type: object
      security:
      - RequireLogin: []
      summary: 获取权限树
      tags:
      - 权限管理
  /sys/permission/delete:
    post:
      consumes:
      - application/json
      description: 根据id删除权限
      parameters:
      - description: 要删除的权限ID
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysPermissionDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 删除权限
      tags:
      - 权限管理
  /sys/permission/page_query:
    post:
      consumes:
      - application/json
      description: 分页查询权限
      parameters:
      - description: 权限筛选条件
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysPermissionPageQueryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/api.PageData'
                  - properties:
                      List:
                        $ref: '#/definitions/models.SysPermission'
                    type: object
              type: object
      security:
      - RequireLogin: []
      summary: 分页查询权限
      tags:
      - 权限管理
  /sys/permission/update:
    post:
      consumes:
      - application/json
      description: 根据id修改权限信息
      parameters:
      - description: 权限信息
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysPermissionUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 修改权限
      tags:
      - 权限管理
...
securityDefinitions:
  RequireLogin:
    in: header
    name: token
    type: apiKey
swagger: "2.0"